requirements.txt
# microwakeword_trainer - Wake Word Training Framework
# Version 2.0 - GPU-Accelerated Training Pipeline

# Core ML Frameworks
python==3.10
tensorflow>=2.16
ai-edge-litert
pymicro-features>=0.1
numpy>=1.26
scipy
pyyaml
mmap_ninja
datasets>=2.14
audiomentations
audio_metadata
webrtcvad-wheels
absl-py

# Performance (v2.0)
cupy-cuda12x>=13.0
pyarrow>=15.0
numba>=0.58

# Optional
speechbrain>=1.0.0
transformers>=4.40.0
scikit-learn>=1.4.0
optuna>=3.6.0
matplotlib
seaborn
